+++ mkidpipeline/hdf/parsebin.pyx
@@ -62,11 +62,11 @@ def parse(file,_n=0):
 
     # Creating pointers to memory bocks that the binlib.c code will fill
     n = int(max(os.stat(file).st_size/8, _n))
-    baseline   = np.empty(n, dtype=int)
-    wavelength = np.empty(n, dtype=float)
+    baseline   = np.empty(n, dtype=np.int)
+    wavelength = np.empty(n, dtype=np.float32)
     timestamp  = np.empty(n, dtype=np.uint64)
-    y = np.empty(n, dtype=np.int)
-    x = np.empty(n, dtype=np.int)
+    y = np.empty(n, dtype=np.uint32)
+    x = np.empty(n, dtype=np.uint32)
     roachnum = np.empty(n, dtype=np.uint32)
 
     # Calling parsebin from binlib.c
